Read, display and output descriptions.
Add support for splitting input lines by `IFS` and using the last part as a
description. The description is not searchable and serves the purpose of
providing context to choices without affecting the search.

Searching the following input with the search string `"Time"` without
support for descriptions would discriminate the `Time` choice since it is
longer than `time 1` and `time 2`. With description support they will all be
treated the same.

```
time 1
time 2
Time 3-ruby-2-1-0
```