Use the sha of the actual release tarball.
The sha of the tarball generated by `make distcheck` is different on every
run because of the file modification time of files in the archive.

Downloading the actual release tarball and using the sha for that in the
Homebrew formula and AUR PKGBUILD ensures that they contain the correct sha.