Allow users to pick the current input query
In some situations users want to pick a choice that is not present in the
list of choices piped to pick.

One example of such a situation is checking out a git revision that is not
in the piped in list:

    git checkout $(git branch | cut -c 3- | pick)

This commit allows the user to pick the current input query enabling them to
pick a commit or a tag by typing it out and pressing Alt-Enter.

This closes #58, and closes #61.