Check that description is not NULL
Certain implementations of `strtok_r` set the context pointer to `NULL` when
the delimiter is not found. This used to cause a segfault when trying to
construct a choice with a `NULL` description.

By explicitly setting description to an empty string if set to `NULL` by
`strtok_r` this segfault is avoided.

This closes #68.