Scroll query horizontally as needed
Previously when a query that wouldn't fit within the number of columns
of the tty was entered, the query input field would be unable to show
the full query.

This commit adds horizontal scrolling to the query input field. When the
cursor is moved outside of the currently visible part of the query, the
query is scrolled to show the previously invisible part.