Reduce the number of scoring states
Assigning a score of 1 to all choices is used to force output of all
choices when the query is empty. Instead modify the loop condition to
check if the query is empty and then bypass the score condition. A
pleasant side-effect of this change is that each score is either 0 or
>0 which implies presence of a score.