Highlight the matched substring
By using the underline capability the matched substring for all choices
is highlighted. This change gives a better notion of the actual match
and allows one to easily spot matched outliers.

The printing of choices is rewritten which removes the need to allocate
a new string for every matched choice and usage of two compat functions.