Add UTF-8 support
Most of the changes involves iterating strings in order to find the next
character since we no longer can assume that every byte is a character.

Another change notable change is the handling of key input: the
refactored `get_key' function ensures a whole Unicode character or
escape sequence is being read.