Reduce flickering
The clearing of the screen introduced in 72239f0 made the screen flicker
for users in tmux. Possibly because tmux might take longer time to draw
than terminal emulators.

Instead of clearing the whole screen before printing the choices and
query, clear from the line below the last choice to the end of the
screen before printing the query. This only can and needs to be done if
the visible choices and the query do not completely fill the screen.