Simplify test input handling
Instead of parsing the input given to the test-runner let printf(1) do
the work. This has several advantages: less code and the sequences
defined in get_key can be used without modification in the test-cases.
With one caveat, new-line can't be sent as is since it's a
input-field-separator. Therefore, \\n is used to include a new-line.

Spaces present in the input-field will be ignored unless they are
preceded by a backslash. This makes the test-cases a bit more easier to
read.