Fix clear bug
The screen was not cleared when scrolling vertically. Running the
following command and pressing page down would leave a dangling choice
which should no longer be visible:

  printf '1\n2\n3\n' | env LINES=3 pick

The solution is to take the current scroll offset into account. While at
it, remove the first condition which is a premature(?) optimization:
only consider clearing the screen if the number of outputted choices is
less than the total number of choices. The current condition is simpler,
clear the screen if there's still room for a extra line. Let's hope this
change does not introduce any flickering while rendering.