Add query option test and fix timing issue
The test spotted a bug related to filter abort introduced in 1ba7398.
The pick-test command enters keyboard input at a very high pace causing
the filtering to be aborted since the input file descriptor contains
unexamined data and 0 mod any number will always equal 0. Therefore,
only abort if at least 50 choices has already been filtered. This is not
a exhaustive solution, consider the following scenario: the number of
choices is greater than 50 and the search is slow, the user inputs any
key that won't affect the query (like choice movement) the filtering
will then be aborted leaving the filtered choices in a bad state. The
correct solution would be to peek at the input file descriptor and only
abort if the first byte is a printable ASCII or UTF-8 character. Such a
solution is complex and this scenario is unlikely to occur.