Optimize min_match
The min_match function calls itself recursively once a match is found in
order to find the minimal matching substring using the given query.
However, if the length of a match is equal to the length of the query
there's no point in continue the recursion since the left-most match is
favored.