Remove usage of save and restore cursor
Usage of cursor save and restore capabilities is non-obvious. Moving the
cursor the appropriate number of lines/columns is easier to follow.

While at it, make the query cursor movement a single operation as
opposed of the current implementation which is linear with the number of
characters in the query.