Optimize min_match
Instead of advancing to the next character in the string to examine,
move passed the first occurrence of the first character in the query.
This is the only place another potential match could reside.

This is an order of magnitude faster on my machine when the number of
choices with a match is large.