Switch to an event driven design
This is probably my last remaining annoyance with pick, the fact that it does a
blocking read of stdin until EOF prior to entering its main loop. If stdin is
slow, this will take time; like running git-log in a huge repository.

The new approach uses poll(2) and can therefore read from stdin as new data
arrives and simultaneously handle user commands. Some notes about the changes:

- get_choices() is made re-entrant. It still allocates a single block of memory
  for the stdin data. Each choice stores an indirect pointer to this piece of
  memory since it could be realloacted invalidating any previously stored
  pointer. This approach is more complex than storing the string in each choice
  but much faster due to less allocations and strdup(3):ing.

- tty_getc() no longer uses the FILE API from libc since it's buffered. Meaning,
  it could potentially read more input stored in its internal buffer during a
  single call to getc(3). If all that data is not consumed during the call to
  get_key(), poll will never flag the underlying file descriptor as non-block
  readable again since the data is already consumed. Issue exhibited by the test
  suite.

To see it in action:

  $ while :; do sleep 1; echo bar; done | pick -q bar