Add COPR instructions for Fedora/RHEL/CentOS users
There is a COPR repo available now.  Many users prefer to avoid COPR
repos for various reasons so the old `wget` instructions are still
fully supported.

This is an important step toward getting included in the official Fedora
repos, so we're getting closer!  Can't put a timeline on inclusion yet
but I'm hoping the next major release will include pick directly.

There are _a lot_ of Fedora packaging standards/rules that I have to go
through and conform too, in addition to the standard process, but it's
coming along.

Closes #312